refactor: split Xcode Cloud client by resource
